perf: adjust the logic of
