+++ src/views/demo/feat/tab-params/index.vue
@@ -1,10 +1,14 @@
 <template>
-  <div class="p-4"> Current Param : {{ params }} </div>
+  <div class="p-4">
+    Current Param : {{ params }}
+    <input />
+  </div>
 </template>
 <script lang="ts">
   import { computed, defineComponent, unref } from 'vue';
   import { useRouter } from 'vue-router';
   export default defineComponent({
+    name: 'TestTab',
     setup() {
       const { currentRoute } = useRouter();
       return {
